requirements.txt
psycopg2-binary==2.9.9
Flask==3.0.2
python-dotenv==1.0.1
gunicorn==21.2.0